move generator.generate_combo to util module
